# docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.LoginInput:
    properties:
      password:
        example: password
        type: string
      role:
        description: admin, member
        example: admin
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  auth.LoginResponse:
    properties:
      token:
        type: string
    type: object
  post.CreatePostInput:
    properties:
      content:
        type: string
      status:
        description: draft, published
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  post.Post:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      slug:
        type: string
      status:
        description: draft, published
        type: string
      title:
        type: string
      updated_at:
        type: string
    required:
    - content
    - title
    type: object
  post.UpdatePostInput:
    properties:
      content:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: RESTful API starterkit using Gin and GORM.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Golang Starterkit API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Simple auth mock endpoint to get a JWT token. Enter "admin" and
        "password" with "admin" role.
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/auth.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Authenticate User & Get JWT Token
      tags:
      - auth
  /posts:
    get:
      consumes:
      - application/json
      description: Get a list of all blog posts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/post.Post'
              type: array
            type: object
      summary: List all blog posts
      tags:
      - posts
    post:
      consumes:
      - application/json
      description: Create a blog post with the given details
      parameters:
      - description: Create Post Input
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/post.CreatePostInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              $ref: '#/definitions/post.Post'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new blog post
      tags:
      - posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Remove a blog post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a blog post
      tags:
      - posts
    get:
      consumes:
      - application/json
      description: Get detail of a blog post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/post.Post'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single blog post
      tags:
      - posts
    put:
      consumes:
      - application/json
      description: Update details of a blog post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Post Input
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/post.UpdatePostInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/post.Post'
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing blog post
      tags:
      - posts
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"